Update hierarchy.mdx with an image of the asset panel
The Assets Panel section of this page has no image to show what it is, unlike the other sections.

I think it should be added, would be really helpful to a total Rive noob like me. For a moment, I was looking around for it wondering what is being talked about.